Begin Session Description class
